Initial data improved

+++ catalogue/app/initial_data.py
@@ -9,7 +9,6 @@ from pathlib import Path
 
 import requests
 from slugify import slugify
-from sqlalchemy import MetaData
 
 from app import crud, models, schemas
 from app.config import settings
@@ -18,6 +17,7 @@ from app.messages import set_logging_disabled
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
+
 
 class bcolors:
     HEADER = "\033[95m"
@@ -57,7 +57,7 @@ def get_snapshots(origin, dest):
             f"{fol}/{file}" for file in os.listdir(snapshots_folder) if file.endswith(('.png', '.jpg', '.jpeg', '.tiff', '.bmp', '.gif'))]
         copy_tree(snapshots_folder, static_snapshots_folder)
 
-        print(snapshots)
+        # print(snapshots)
         return snapshots
     return []
 
@@ -69,40 +69,49 @@ def get_logotype(logotype_path, origin, dest):
     ori = str(origin) + "/" + file_path
     dst = f"{dest}/logotype{file_extension}"
     move_file(ori, dst)
-    print(dst)
+    # print(dst)
     return dst.replace("/app", "")
 
 
 async def create_interlinker(db, metadata_path, software=False, externalsoftware=False, knowledge=False, externalknowledge=False):
     error = False
     ####################
-    # COMMON
+    # COMMON for all interlinker types
     ####################
     str_metadata_path = str(metadata_path)
     with open(str_metadata_path) as json_file:
         data = json.load(json_file)
 
+    # get the english name of the interlinker for finding it
     name = data["name_translations"]["en"]
     print(f"\n{bcolors.OKBLUE}Processing {bcolors.ENDC}{bcolors.BOLD}{name}{bcolors.ENDC}")
 
+    # get the existing interlinker
     existing_interlinker = await crud.interlinker.get_by_name(db=db, name=name)
     if (existing_interlinker):
         print(f"\t{bcolors.WARNING}Already in the database{bcolors.ENDC}. UPDATING")
 
-    # parent folder where metadata.json is located
+    # Get the parent folder object where the metadata.json is located
     folder = metadata_path.parents[0]
     slug = slugify(name)
     str_static_path = f'/app/static/{slug}'
+
+    # Copy the snapshots and the logotype
     data["snapshots"] = get_snapshots(folder, str_static_path)
     data["logotype"] = get_logotype(
         data["logotype"], folder, str_static_path) if "logotype" in data else None
 
-    # get instructions file contents if IS FILE PATH
-    for key, value in data["instructions_translations"].items():
-        if not "http" in value:
-            filename = path_leaf(value)
+    # Get the instructions file contents (normally is a README file, but it can be a pointer to a URI. 
+    # In that case, the URI is stored, there is no need to read the content of the file)
+    for LANGUAGE_KEY, FILE_NAME_OR_URL in data["instructions_translations"].items():
+        if not "http" in FILE_NAME_OR_URL:
+            filename = path_leaf(FILE_NAME_OR_URL)
             with open(str(folder) + "/" + filename, 'r') as f:
-                data["instructions_translations"][key] = f.read()
+                data["instructions_translations"][LANGUAGE_KEY] = f.read()
+
+    ###################################
+    # ONLY FOR SOFTWARE INTERLINKERS
+    ###################################
 
     if software:
         # set nature
@@ -119,13 +128,17 @@ async def create_interlinker(db, metadata_path, software=False, externalsoftware
                 db_obj=existing_interlinker,
                 obj_in=schemas.SoftwareInterlinkerPatch(**data),
             )
+            print(f"\t{bcolors.OKGREEN}Updated successfully!{bcolors.ENDC}")
         else:
             interlinker = await crud.interlinker.create(
                 db=db,
                 interlinker=schemas.SoftwareInterlinkerCreate(**data),
             )
+            print(f"\t{bcolors.OKGREEN}Created successfully!{bcolors.ENDC}")
 
-
+    ###################################
+    # ONLY FOR EXTERNAL KNOWLEDGE INTERLINKERS
+    ###################################
     if externalknowledge:
         data["nature"] = "externalknowledgeinterlinker"
 
@@ -135,15 +148,20 @@ async def create_interlinker(db, metadata_path, software=False, externalsoftware
                 db_obj=existing_interlinker,
                 obj_in=schemas.ExternalKnowledgeInterlinkerPatch(**data),
             )
+            print(f"\t{bcolors.OKGREEN}Updated successfully!{bcolors.ENDC}")
         else:
             interlinker = await crud.interlinker.create(
-            db=db,
-            interlinker=schemas.ExternalKnowledgeInterlinkerCreate(**data),
-        )
+                db=db,
+                interlinker=schemas.ExternalKnowledgeInterlinkerCreate(**data),
+            )
+            print(f"\t{bcolors.OKGREEN}Created successfully!{bcolors.ENDC}")
 
-        
+    ###################################
+    # ONLY FOR EXTERNAL SOFTWARE INTERLINKERS
+    ###################################
+
     if externalsoftware:
-         # set nature
+        # set nature
         data["nature"] = "externalsoftwareinterlinker"
 
         if existing_interlinker:
@@ -152,13 +170,23 @@ async def create_interlinker(db, metadata_path, software=False, externalsoftware
                 db_obj=existing_interlinker,
                 obj_in=schemas.ExternalSoftwareInterlinkerPatch(**data),
             )
+            print(f"\t{bcolors.OKGREEN}Updated successfully!{bcolors.ENDC}")
         else:
             interlinker = await crud.interlinker.create(
                 db=db,
                 interlinker=schemas.ExternalSoftwareInterlinkerCreate(**data),
             )
+            print(f"\t{bcolors.OKGREEN}Created successfully!{bcolors.ENDC}")
 
-    # TODO: update files and remove old ones 
+    ###################################
+    # ONLY FOR KNOWLEDGE INTERLINKERS
+    ###################################
+    # Knowledge interlinkers are the most complex type of interlinkers to be updated. As the knowledge interlinkers point to an asset managed by a software interlinkers,
+    # there is no (or easy) way to know if the asset used to create the interlinker has changed.
+    # For example, we are using a pptx to build the interlinker Business Model Canvas (check seed/interlinkers/knowledge/Business Model Canvas).
+    # To "Update" them, we make a DELETE request to the interlinker (with the genesis asset id of the knowledge interlinker) and we create it again with the pptx in the folder.
+    # This way, we do not care if it has changed or not. It is not the optimal way, but the easiest to maintain.
+
     if knowledge:
         existing_interlinker: models.KnowledgeInterlinker
         # get file contents in file and send to the software interlinker
@@ -172,59 +200,72 @@ async def create_interlinker(db, metadata_path, software=False, externalsoftware
             print(f"\tis {service} supported knowledge interlinker")
 
             genesis_asset_id_translations = {}
-            for key, value in data["file_translations"].items():
-                filename = path_leaf(value)
+            for LANGUAGE_KEY, COMPLE_FILE_URL in data["file_translations"].items():
+                # Get only the name of the file (we already know the folder where it is located)
+                filename = path_leaf(COMPLE_FILE_URL)
+                
+
+                # Delete existent asset in the interlinker for the given language key
+                existing_asset_id = existing_interlinker.genesis_asset_id_translations.get(LANGUAGE_KEY, None)
+                if existing_asset_id:
+                    URL = f"http://{softwareinterlinker.service_name}{softwareinterlinker.api_path}/{existing_asset_id}"
+                    response = requests.delete(URL, headers={
+                        "Authorization": settings.BACKEND_SECRET
+                    })
+                    if (response.status_code < 200 or response.status_code >= 300) and response.status_code != 404:
+                        raise Exception(response.json())
+
+                # Create a new asset with the content of the file in file_translations
                 short_filename, file_extension = os.path.splitext(filename)
 
-                # Delete existent asset in the interlinker
-                existing_asset_id = existing_interlinker.genesis_asset_id_translations[key]
-                URL = f"http://{softwareinterlinker.service_name}{softwareinterlinker.api_path}/{existing_asset_id}"
-                response = requests.delete(URL, headers={
-                    "Authorization": settings.BACKEND_SECRET
-                })
-                if response.status_code < 200 or response.status_code >= 300:
-                    raise Exception(response.json())
-
-                # Create a new one
+                ## if the file is json, read the content and send it in post data as json
                 if "json" in file_extension:
                     with open(str(folder) + "/" + filename, 'r') as f:
                         response = requests.post(
-                            f"http://{service}/assets", data=f.read()).json()
+                            f"http://{service}/assets", data=f.read())
+                
+                ## if the file is a file (like pdf, pptx...), read the content and send it in post data as file
                 else:
-                    
                     filedata = open(str(folder) + "/" + filename, "rb").read()
-                    name_for_file = data["name_translations"][key]
+                    name_for_file = data["name_translations"][LANGUAGE_KEY]
                     files_data = {
                         'file': (name_for_file + file_extension, filedata)}
                     response = requests.post(
-                        f"http://{service}/assets", files=files_data).json()
+                        f"http://{service}/assets", files=files_data)
 
-                print(f"{key} ANSWER FOR {service}")
-                print(response)
-                data["softwareinterlinker_id"] = softwareinterlinker.id
-                genesis_asset_id_translations[key] = response["id"] if "id" in response else response["_id"]
-            
+                if response.status_code < 200 or response.status_code >= 300:
+                    raise Exception(response.json())
+                
+                response_data = response.json()
+                genesis_asset_id_translations[LANGUAGE_KEY] = response_data["id"] if "id" in response_data else response_data["_id"]
+                print(f"Asset for the {LANGUAGE_KEY} has been created and added in the genesis_asset_id_translations")
+
+            # Last fields needed for the knowledge interlinker
+            data["softwareinterlinker_id"] = softwareinterlinker.id
             data["genesis_asset_id_translations"] = genesis_asset_id_translations
             if existing_interlinker:
-                interlinker : models.KnowledgeInterlinker = await crud.interlinker.update(
+                interlinker: models.KnowledgeInterlinker = await crud.interlinker.update(
                     db=db,
                     db_obj=existing_interlinker,
                     obj_in=schemas.KnowledgeInterlinkerPatch(**data),
                 )
+
+                # Comprobation that the genesis asset ids have been updated
                 if interlinker.genesis_asset_id_translations != genesis_asset_id_translations:
-                    raise Exception(f"{interlinker.genesis_asset_id_translations} no es igual a {genesis_asset_id_translations}")
+                    raise Exception(
+                        f"{interlinker.genesis_asset_id_translations} not equal to {genesis_asset_id_translations}")
+                print(f"\t{bcolors.OKGREEN}Updated successfully!{bcolors.ENDC}")
             else:
-                interlinker : models.KnowledgeInterlinker = await crud.interlinker.create(
+                interlinker: models.KnowledgeInterlinker = await crud.interlinker.create(
                     db=db,
                     interlinker=schemas.KnowledgeInterlinkerCreate(**data),
                 )
+                print(f"\t{bcolors.OKGREEN}Created successfully!{bcolors.ENDC}")
 
         except Exception as e:
             error = True
             print(f"\t{bcolors.FAIL}{str(e)}{bcolors.ENDC}")
-        if not error:
-            verb = "Created" if not existing_interlinker else "Updated"
-            print(f"\t{bcolors.OKGREEN}{verb} successfully!{bcolors.ENDC}")
+
 
 async def create_problemprofile(db, problem):
     id = problem["id"]
@@ -259,15 +300,15 @@ async def create_coproductionschema(db, schema_data):
         )
     else:
         SCHEMA = await crud.coproductionschema.create(
-        db=db,
-        obj_in=schemas.CoproductionSchemaCreate(
-            **schema_data, is_public=True
+            db=db,
+            obj_in=schemas.CoproductionSchemaCreate(
+                **schema_data, is_public=True
+            )
         )
-    )
 
     print(f"{bcolors.OKBLUE}## Processing {bcolors.ENDC}{name}")
     items_resume = {}
-    
+
     phase_data: dict
     touched_phases = []
     for phase_data in schema_data["phases"]:
@@ -295,10 +336,12 @@ async def create_coproductionschema(db, schema_data):
         touched_objectives = []
         for objective_data in phase_data["objectives"]:
             if objective_in_db := await crud.treeitems.get_by_names(db=db, name_translations=objective_data["name_translations"], parent_id=db_phase.id):
-                print(f"{bcolors.WARNING}Updating existing objective {objective_in_db.name}{bcolors.ENDC}")
+                print(
+                    f"{bcolors.WARNING}Updating existing objective {objective_in_db.name}{bcolors.ENDC}")
                 db_objective = await crud.treeitems.update(db=db, db_obj=objective_in_db, obj_in=schemas.TreeItemPatch(**objective_data))
             else:
-                objective_name = objective_data.get("name_translations", {}).get("en", "undefined")
+                objective_name = objective_data.get(
+                    "name_translations", {}).get("en", "undefined")
                 print(f"{bcolors.OKGREEN}Creating {phase_name} - {objective_name}{bcolors.ENDC}")
                 db_objective = await crud.treeitems.create(
                     db=db,
@@ -318,11 +361,14 @@ async def create_coproductionschema(db, schema_data):
             touched_tasks = []
             for task_data in objective_data["tasks"]:
                 if task_in_db := await crud.treeitems.get_by_names(db=db, name_translations=task_data["name_translations"], parent_id=db_objective.id):
-                    print(f"{bcolors.WARNING}Updating existing task {task_in_db.name}{bcolors.ENDC}")
+                    print(
+                        f"{bcolors.WARNING}Updating existing task {task_in_db.name}{bcolors.ENDC}")
                     db_task = await crud.treeitems.update(db=db, db_obj=task_in_db, obj_in=schemas.TreeItemPatch(**task_data))
                 else:
-                    task_name = task_data.get("name_translations", {}).get("en", "undefined")
-                    print(f"{bcolors.OKGREEN}Creating {phase_name} - {objective_name} - {task_name}{bcolors.ENDC}")
+                    task_name = task_data.get(
+                        "name_translations", {}).get("en", "undefined")
+                    print(
+                        f"{bcolors.OKGREEN}Creating {phase_name} - {objective_name} - {task_name}{bcolors.ENDC}")
                     db_task = await crud.treeitems.create(
                         db=db,
                         obj_in=schemas.TreeItemCreate(
@@ -333,14 +379,15 @@ async def create_coproductionschema(db, schema_data):
                     )
                 touched_tasks.append(db_task)
 
-                sum = list(task_data["problemprofiles"]) + list(objective_data["problemprofiles"])
+                sum = list(task_data["problemprofiles"]) + \
+                    list(objective_data["problemprofiles"])
                 await crud.treeitems.sync_problemprofiles(db=db, treeitem=db_task, problemprofiles=sum, commit=False)
 
                 items_resume[task_data["id"]] = {
                     "db_id": db_task.id,
                     "prerequisites": task_data.get("prerequisites", [])
                 }
-            
+
             for child in db_objective.children:
                 if child not in touched_tasks:
                     print(f"{bcolors.FAIL}Removing task {child.name}{bcolors.ENDC}")
@@ -367,6 +414,7 @@ async def create_coproductionschema(db, schema_data):
             print(f"{bcolors.FAIL}Removing phase {child.name}{bcolors.ENDC}")
             await crud.treeitems.remove(db=db, id=child.id)
 
+
 async def init():
     db = SessionLocal()
     set_logging_disabled(True)
@@ -386,7 +434,7 @@ async def init():
         # create external interlinkers first
         for metadata_path in Path("/app/seed/interlinkers").glob("externalsoftware/**/metadata.json"):
             await create_interlinker(db, metadata_path, externalsoftware=True)
-        
+
         for metadata_path in Path("/app/seed/interlinkers").glob("externalknowledge/**/metadata.json"):
             await create_interlinker(db, metadata_path, externalknowledge=True)
 
@@ -400,7 +448,7 @@ async def init():
 
     except Exception as e:
         raise e
-        
+
     db.close()
 
 if __name__ == "__main__":
@@ -408,4 +456,3 @@ if __name__ == "__main__":
     loop = asyncio.get_event_loop()
     loop.run_until_complete(init())
     logger.info("Initial data created")
-
